REF: Move 2 duplicated util functions to cmtklib/utils.py
