+++ cmp/stages/segmentation/segmentation.py
@@ -23,7 +23,8 @@ import nipype.interfaces.utility as util
 # Own imports
 from cmp.stages.common import Stage
 from cmtklib.interfaces.freesurfer import copyBrainMaskToFreesurfer
-from cmtklib.util import extract_freesurfer_subject_dir, extract_reconall_base_dir
+from cmtklib.util import isavailable, extract_freesurfer_subject_dir,\
+    extract_reconall_base_dir, get_freesurfer_subject_id
 
 
 class SegmentationConfig(HasTraits):
@@ -208,24 +209,6 @@ class SegmentationStage(Stage):
                     # fs_reconall.inputs.hippocampal_subfields_T1 = self.config.segment_hippocampal_subfields
                     # fs_reconall.inputs.brainstem = self.config.segment_brainstem
 
-                    def isavailable(file):
-                        """Check if file is available and return the file if it is.
-
-                        Used for debugging. (duplicated later)
-
-                        Parameters
-                        ----------
-                        file : File
-                            Input file
-
-                        Returns
-                        -------
-                        file : File
-                            Output file
-                        """
-                        # print "T1 is available"
-                        return file
-
                     flow.connect([
                         (inputnode, fs_mriconvert, [
                          (('T1', isavailable), 'in_file')]),
@@ -248,24 +231,6 @@ class SegmentationStage(Stage):
 
                     # fs_reconall.inputs.subjects_dir and fs_reconall.inputs.subject_id set in cmp/pipelines/diffusion/diffusion.py
                     fs_autorecon1.inputs.subjects_dir = self.config.freesurfer_subjects_dir
-
-                    def isavailable(file):
-                        """Check if file is available and return the file if it is.
-
-                        Used for debugging. (Duplicata)
-
-                        Parameters
-                        ----------
-                        file : File
-                            Input file
-
-                        Returns
-                        -------
-                        file : File
-                            Output file
-                        """
-                        # print "Is available"
-                        return file
 
                     flow.connect([
                         (inputnode, fs_mriconvert, [
@@ -355,22 +320,6 @@ class SegmentationStage(Stage):
                     #             (fs_source,fs_mriconvert_nu,[('nu','in_file')])
                     #             ])
 
-                    def get_freesurfer_subject_id(file):
-                        """Extract Freesurfer subject ID from file generated by recon-all.
-
-                        Parameters
-                        ----------
-                        file : str
-                            File generated by recon-all
-
-                        Returns
-                        -------
-                        out : str
-                            Freesurfer subject ID
-                        """
-                        # print("Extract reconall base dir : %s" % file[:-18])
-                        out = file[:-18]
-                        return out
 
                     fs_reconall23 = pe.Node(interface=fs.ReconAll(
                         flags='-no-isrunning -parallel -openmp {}'.format(self.config.number_of_threads)), name='reconall23')
