+++ cmtklib/util.py
@@ -39,6 +39,24 @@ def unicode2str(text):
     return out_str
 
 
+def isavailable(file):
+    """Check if file is available and return the file if it is.
+
+    Used for debugging.
+
+    Parameters
+    ----------
+    file : File
+        Input file
+
+    Returns
+    -------
+    file : File
+        Output file
+    """
+    return file
+
+
 class BColors:
     """Utility class for color unicode."""
     HEADER = '\033[95m'
@@ -385,6 +403,23 @@ def extract_freesurfer_subject_dir(reconall_report, local_output_dir=None, debug
             cnt += 1
 
     return fs_subject_dir
+
+
+def get_freesurfer_subject_id(file):
+    """Extract Freesurfer subject ID from file generated by recon-all.
+
+    Parameters
+    ----------
+    file : str
+        File generated by recon-all
+
+    Returns
+    -------
+    out : str
+        Freesurfer subject ID
+    """
+    out = file[:-18]
+    return out
 
 
 def extract_reconall_base_dir(file):
